langchain_multistage_taxonomy/requirements.txt
# Dependencies for LangGraph legal case classification workflow
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.2.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0 